FinAssist - A making of Finance management tool
This is an application for solving problems faced by small business
owners such as restaurants owners, kirana shops, tailor shops etc. The
interface and functionality are only essential and clear to be most
helpful to people who are likely not too tech savvy and possibly
illiterate in some cases.

The app helps them to manage their financials in an infographic way and
lets them keep track of the finances,